+++ stack.go
@@ -163,7 +163,7 @@ func funcname(name string) string {
 
 // StackFormat formats the provided stack trace as text.
 //
-// For each frame in the stack, separated by \n:
+// For each frame in the stack, ended by \n:
 //
 //	%s	  lists the source file
 //	%d    lists the source line
@@ -177,9 +177,8 @@ func funcname(name string) string {
 //	%+v   lists the full function name and full compile-time path of the source file
 //	      with the source line, separated by \n\t
 //	      (<funcname>\n\t<path>:<line>)
-func StackFormat(w io.Writer, format string, s []uintptr) (int, E) {
-	n, err := fmt.Fprintf(w, format, stack(s))
-	return n, WithStack(err)
+func StackFormat(format string, s []uintptr) string {
+	return fmt.Sprintf(format, stack(s))
 }
 
 // StackMarshalJSON marshals the provided stack trace as JSON.